[BACKLOG-40174] - support custom icons in F-O-S dialogs
In both left tree and right pane
